Fix EqProcessor rounding, add restricted pointers and noise-shaped dither, consolidate wire-scale constants, enable -Wconversion on ae_usb
- process16/process24 now round-to-nearest instead of truncating, matching process32's existing quantize discipline (dead in this app today, but the shared engine has other consumers)
- __restrict on hot DSP/USB pointer params (measured win on processToDouble)
- wire_scale.h centralizes the previously duplicated 32767.0/8388607.0/2147483647.0 constants
- New ae::NoiseShapedQuantizer: first-order noise-shaped TPDF dither for 16-bit Reference EQ output, kept separate from TpdfQuantizer so its flat-dither oracle stays untouched
- ae_usb now builds clean under -Wall -Wextra -Wconversion; every narrowing made explicit, no real bugs found
- DsdDecoder reserves its worst-case packing buffer once instead of resizing every read()
- dsp_null_test: new coverage for the rounding fix and NoiseShapedQuantizer, ARM EQ check tightened to exact equality now that the NEON path is FMA-free (unverified on real ARM hardware)
- dsp_bench: new NoiseShapedQuantizer benchmark case

+++ core/include/core/dsp/audio_convert.h
@@ -19,8 +19,10 @@ public:
     // Returns uniform random in [-0.5, +0.5]
     inline float next() {
         state = state * 1664525u + 1013904223u;
-        // Convert upper bits to float in [0, 1), subtract 0.5
-        return (state >> 8) * (1.0f / 16777216.0f) - 0.5f;
+        // Convert upper bits to float in [0, 1), subtract 0.5. state >> 8 is
+        // a 24-bit value (0..16777215), exactly representable in float's
+        // 24-bit mantissa — the cast is exact, not a truncation.
+        return static_cast<float>(state >> 8) * (1.0f / 16777216.0f) - 0.5f;
     }
 
     // TPDF: sum of two uniform randoms -> triangular distribution in [-1, +1]
